+++ sources/Panel/src/Generator/InterfacePanel.h
@@ -1,4 +1,5 @@
 #pragma once
+#include "Utils/Buffer.h"
 
 
 class Interface
@@ -7,7 +8,7 @@ public:
     /// Принять и обработать данные от ПЛИС
     static void ProcessDataFPGA();
     /// Заслать в генератор данные
-    static void Send(const uint8 *buffer, uint16 size);
+    static void Send(const uint8 *buffer, uint size);
 
     static void Send(const Buffer &buffer);
     /// Принять numBytes байт от ПЛИС и выполнить их
